Start of Strength/projectiles.
Projectiles are somewhat added.
Basic controller which goes towards mouse pointer, strength code is written player side but the projectiles need access to that variable.

Noted bugs on "Delete Entity" which need to be addressed i.e. The UID list is a stack and not an array, meaning that the secondary list will attempt to connect random ID's on a list with grows iteratively.