Doors, Corridors and a small UI fix
Rooms are now linked through corridors with doors that can be opened using the "." key and closed using the "," key.
Each corridor will need a door with a wall, the wall is basically a tile that will be used in place of the door if the corridor so ideally it should match the rest of the room walls.
A small issue where the progress bars for throwing had inconsistent background and wrong anchor points was fixed.  Canvas tag was also corrected from finish to a unique tag made specifically for charge bars.