Started Weapon on player
WeaponController is a bit of a mess and will require additional work.

Player now holds a reference to whatever weapon is held, alongside code to allow the player to pick up weapons.

Weapon has an accurate collision detection for interacting with the player despite their collision boxes not interacting, allowing the player to move through projectiles.