Fixed Weapon pickup
Weapons are now picked up by the player though there is one known bug which will need to be addressed.
>Clones are treated as base weapons

Currently weapons are placed at Vector3.zero, to keep them as reference. Additional "Pickups" will be deleted. There needs to be a "IsProjectile" check which does not interfere with the held "Weapon" projectile.