Weapons 1.0
>Spawn weapons on crosshair with L

Debug properly in the future when we have more weapons but it seems to work properly with a single weapon option.

Can throw a weapon on each hand and can't throw again until you pick it up. Easy to change if it needs to be "Ammo" rather than an IfElse

Most of the work was limited to WeaponController and Player - The latter of which interacting with the "Stored" weapon. It still uses prefabs despite a weapon reference being stored at Vector3.zero so that needs to be addressed.