Minor changes
Added PriorityChar_Manager in order to increase efficiency when grabbing the player. Finding an object by tag is not efficient and doing it on every enemy entity could slow down the game when the count is increased.

Future important characters will be added here, when referenced a lot. This also allows AI to work more intelligently by efficiently getting the positions of a multitude of gameobjects without iterating through the scene.