Started interaction
Text Mesh Pro seems to be deprecated so I've copied an Icon from the main menu, since that seems to work fine.

Starting player with the burger to make an interactable NPC. It will now create a radius around itself whenever you select any entity with the class which broadcasts its "InteractionRange". Thusfar it does nothing.

In the next update it should spawn an icon as an entity, telling the player to interact with the object.