Basic room generation finished
Rooms are generated on a grid and connected to the neighbouring rooms.
The rooms use 2 prefabs lists, one for normal lists and another one for boss rooms.
Boss rooms yet need to have their game logic implemented (currently the rooms is a normal room)
Doors linking rooms all have a connection point that is 3 blocks away from the specified door tile that the corridor generation will try to connect to, the room figures out which direction the door is on its own.
Currently all rooms need a door in each direction.
Obstacle/prop support is yet to be implemented.