Porting work
Been working on Grads in games, in the name of efficiency I designed systems which could work in multiple games. Currently porting over the inventory system and I have a basic system that limits player control when active but is only bound to the player, as of this time.

This is a checkpoint for any errors that may occur.

See. UI - This will be retrofitted into working as a manager.