Sound effects
Sound effects on deaths are a bit glitchy still - I'm working on it. But this seems like a good checkpoint to stop.

Added a soundmanager which can be given to entities to allow for a randomized "On movement" or "On hit" sound

Currently instantiates an empty game object to allow the sound to continue after death but this will need to be added to a separate entity list to allow for the sound to be deleted.