Final Dash until feedback
Chances are, I'll add an incrementing cooldown (Use Dash within 0.25 seconds of cooldown end increases cooldown by 0.1, or something)
but for now, this seems to be fine.

Dash is a separate thread as it overhauls the "Move" function (Which is a single use function per frame) to act over a multitude of frames, locking the function out of use until its end.

Need feedback for more ease of use as, as the coder, I am biased towards my own work.