Map fixes, props and sprite recolouring
Placement of U shaped walls was fixed.
Props that are not near walls are now supported.
Some objects that were not casting or receiving shadows were fixed.
A recolour shader was added, each entity now has 3 colours to modify, if the colourable shader is given to the entity, those 3 colours will replace the red, green and blue colours of the sprite.
In order to use this properly, the sprites will need to be recoloured to only use red, green and blue.
Darker shades of these colours will produce darker shades of the selected colour.