0.95 Update
Mostly finished with readjusting weapons and dash.

Weapons have a few issues, such as the charging bars not displaying properly, but dashes have been updated.

Weapon chargeup has been disabled for this version but will be available in the next push.

Do note that the weapon 'throw' is now shown via the animation so I will need help adjusting that value.