Entity flashing and minor fixes
Player (and the bed) were made smaller to make the dodging easier.
Player and the enemies now flash different colours when hit, during the flashing, the entity is invincible.
Doors are no longer drawn in front of the player.
Doors should no longer push the player out of the map.
Dash speed, distance and animation enter adjusted.
Door size adjusted.