Prop damage, enemy and boss list, death screen, blood splashes and reflections
Props can be damaged, a visual flash happens just the way it does for entities.
Boss and enemy lists are now a part of the map instance, change them to modify which enemies appear on the map.
Death screen now shows when the player is killed.
Blood splashes now show when hit, each character has their own list of splashes, a mirror shader was added to add reflections to splashes, it will reflect anything that has a sprite renderer and a reflection prefab added to it.