Minor quality of life changes
Mostly looking into why the new weapon spawns don't get picked up. Identified it as the DroppedStateUpdate collision detection not working properly. Not sure of the issue but I'm now behind on Game in a Day so I will delay updating this until then.

Outside of that: Added user defined interaction radius's for the vending machine and went over the weapon code, cleaning up the logic in some areas.

Logic appears mostly fine, it's just the mentioned collision code not enabling.