Added prop rarity, fixed walls and linking rooms
The map now has 6 lists for props, both near wall props and regular props have been given 3 rarities, common, uncommon and rare, each rarity has a chance of appearing that is adjusted in the Globals file, the chance is out of the sum of the chances (if the chances add up to 100, it will be out of 100).

The rooms are now linked to doors going in the opposite direction (east to west etc.) instead of linking the closest ones to prevent them from overlapping the rooms.