New Maps and Random floor and wall tiles
I have added 2 more tile maps, only changed the floor tiles so I could tell if changing maps works.

I also added some extra tiles to the map generation to have random wall and floor cracks show up.

Now every time you move to a dungeon, the map you end up on is randomised. Each tile has an equal chance of showing up so if you want a tile to be more common, just add more of the same tile to the tile list in the map prefab.

If you want to edit the maps, use the map prefabs and not the maps in the scenes, adding new maps will require you to add the currently opened scene to unity so the build gets the map, adding the name of the scene to the globals file and incrementing the constant in the globals file, then you can just simply give the scene a new map prefab and it should work.